spawnStreamingLoop(): add a restart cooldown
belacoder will no longer attempt to restart its pipeline on errors
and this will become belaUI's responsibility. However, we'll start
an infinite loop of USB & UVC driver errors if we attempt to do so
immediately after the pipeline stalls due to changing the input
resolution to a Camlink 4K on a Jetson Nano 4GB while streaming
from it.

Delaying belacoder restarts by 2 seconds seems to avoid this issue
very reliably on my setup, but it's an ugly hack and it would be
nice to get a proper fix in the future. The USB hub used on the
Nano 4GB seems at least partly responsible for this, as the Jetson
Nano 2GB, which doesn't use an USB hub for the USB 3.0 port, is
unaffected.

TL;DR: use a Jetson Nano 2GB, don't change resolution while
streaming with a Camlink 4K