Wifi: match known connections to network interfaces by MAC adddress
This is a temporary proof-of-concept. The related WiFi management code
will be further modified and refactored

Matching connections to an ifname as done previously isn't a good idea,
because the name of the interface can change between reboots, if
additional WiFi adapters are plugged in, if there's an USB glitch, etc.
Ubuntu 18.04 as distributed for Jetson Nano doesn't use persistent
interface names. This was also making the WiFi manager ask for the
password and create new NM connections when trying to connect to
networks already known to NM.

Furthermore, this was redundant. NetworkManager already associates the
connections with the WiFi adapter's MAC address.