Audio: wait for the selected card if unavailable rather than automatically switching to a fallback
Imagine the following scenario: you normally use an external USB mic, but
you try to start the stream before turning it on. If belaUI switches
to the fallback, you then have to stop the stream, turn on the mic,
wait for it to be detected, and start the stream again.

After this change, belaUI will just wait for the mic to be connected
before starting the stream, so you just have to turn the mic on