Faster Wifi connection status update on disconnection
We found that once a Wifi connection fails e.g. due to going out
of range or because the hotspot goes down, then the RUNNING flag
is immediately removed from the interface, while nmcli still shows
an established connection and the interface still has an IP
address configured for around 10 seconds longer

This patch uses a missing RUNNING flag to detect disconnections
earlier. Additionally, we trigger wifiScheduleScanUpdates() on
disconnection so that the network list can be automatically updated
to remove the AP that's now no longer in range. This seems to take
around 10-15 seconds, hence the additional polling added to
wifiScheduleScanUpdates()