Make sendInitialStatus() synchronous
This avoids potential races between sendInitialStatus() and
other functionality using broadcastMsg(), ensuring that
1) the full initial config / state is received by the client
   before any broadcasted updates, and
2) a newly connected client also doesn't miss any broadcasted
   messages generated between authentication and the full
   initial status having been sent

The tradeoff is that the pipelines are now no longer updated when
a new client connection authenticates. This is expected to have
negligible impact, as 1) creating, removing or updating pipeline
files isn't part of the regular workflow, and because 2) now the
experience of using local connections matches that of using the
cloud remote, i.e. restarting belaUI is needed after modifying the
pipeline files